garmin/requirements.txt
garminconnect>=0.2.22
fitparse>=1.2.0
python-dotenv>=1.0.0
